+++ src/main/java/io/github/H0n0riuss/calculations/frequency/BoxPlot.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 
 public class BoxPlot {
 
-    public BoxPlotModel calculateBoxplot(float[] data) throws ArithmeticException {
+    public BoxPlotModel calculateBoxplot(double[] data) throws ArithmeticException {
         if (data.length == 0) throw new ArithmeticException("Contains no data.");
 
         Arrays.sort(data);
@@ -20,9 +20,9 @@ public class BoxPlot {
                 .build();
     }
 
-    private float calculateQuantile(float[] data, float percentage) { //TODO wie wird das Quantil wirklich gebildet (ist es das nächste an 25/75% oder aus der Summe)
+    private double calculateQuantile(double[] data, double percentage) { //TODO wie wird das Quantil wirklich gebildet (ist es das nächste an 25/75% oder aus der Summe)
         var test = (data.length - 1) * percentage;
-        var index = Math.round(test);
+        var index = (int) Math.round(test);
         return data[index];
     }
 }
